+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 
 ---
 
+## [1.0.2] - 2026-02-09
+### Added
+- [Temperature] new struct for temperature units
+- [Volume] new struct for physical volume size units
+- [Length] operator * Area to Volume
+- [Area] operator * Lenght to Volume
+- [DataVolume] new struct for data volume size units
+- [Speed] new struct for speed units
+- [Frequency] new sturct for frequency units
+- [Angle] new struct for angle units
+- [Pressure] new struct for pressure units
+- [Calculators] Hydrostatics to calculate Pressue by depth in fresh/salt water
+
 ## [1.0.1] - 2025-12-21
 ### Added
 - [GeoLocation] move from Library